services/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
pymysql==1.1.0
cryptography==41.0.8
redis==5.0.1
influxdb-client==1.39.0
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
websockets==12.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1 